+++ index.html
@@ -233,3 +233,5 @@
 </body>
 </html>
 <script src="js/main.js"></script>
+<script type=module src="js/modules/getRandom.js"></script>
+<script type=module src="js/modules/generateOfferSummary.js"></script>
